+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Refresh playlist to save video ids
 - Ability to get videos with no or specific "status" field
 - Ability to set "status" fields for videos
+- Ability to read specific fields for videos from loaded json